requirements.txt
flask==2.0.1
pymongo==4.0.1
flask-cors==3.0.10
cryptography==36.0.1
werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
python-dotenv==1.0.0
yagmail
keyring 